+++ CHANGELOG.md
@@ -8,6 +8,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Calendar Versioning](https://calver.org).
 
 
+## [2023.10.1] - 2023-11-04
+
+### Changed
+- [CSE] Added &lt;FCNT> as child resource type for &lt;NOD>. Thanks to @BobFIVE for the contribution.
+
+### Security
+- [MISC] Bumped referenced library werkzeug to version 3.0.1
+
+
 ## [2023.10] - 2023-10-23
 
 ### Added